+++ procrustus_indexer/__init__.py
@@ -1,0 +1,2 @@
+from .indexer import Indexer
+from .indexer_builder import build_indexer
