+++ procrustus_indexer/indexer.py
@@ -1,0 +1,83 @@
+from elasticsearch import Elasticsearch
+from elasticsearch.helpers import bulk
+
+from procrustus_indexer.parsers import Parser
+
+
+class Indexer:
+    es: Elasticsearch = None
+    config: dict
+    index_name: str
+    parser: Parser
+
+    def __init__(self, es: Elasticsearch, config: dict, parser: Parser, index_name: str):
+        self.es = es
+        self.config = config
+        self.index_name = index_name
+        self.parser = parser
+
+
+    def create_mapping(self, overwrite: bool = False) -> dict:
+        """
+        Create the elasticsearch index mapping according to config and return resulting dict.
+        :return:
+        """
+        if overwrite:
+            self.es.indices.delete(index=self.index_name, ignore=[400, 404])
+
+        properties = {}
+        for facet_name in self.config['index']['facet'].keys():
+            facet = self.config["index"]["facet"][facet_name]
+            property_type = facet.get('type', 'text')
+            if property_type == 'text':
+                properties[facet_name] = {
+                    'type': 'text',
+                    'fields': {
+                        'keyword': {
+                            'type': 'keyword',
+                            'ignore_above': 256
+                        },
+                    }
+                }
+            elif property_type == 'keyword':
+                properties[facet_name] = {
+                    'type': 'keyword',
+                }
+            elif property_type == 'number':
+                properties[facet_name] = {
+                    'type': 'integer',
+                }
+            elif property_type == 'date':
+                properties[facet_name] = {
+                    'type': 'date',
+                }
+
+        mappings = {
+            'properties': properties
+        }
+
+        settings = {
+            'number_of_shards': 1,
+            'number_of_replicas': 0
+        }
+
+        self.es.indices.create(index=self.index_name, mappings=mappings, settings=settings)
+        return mappings
+
+
+    def import_files(self, files: list[str]):
+        """
+        Import files into an elasticsearch index based on the given config.
+        :param files: list of files to import
+        :param index: Elasticsearch index
+        :return:
+        """
+        es = Elasticsearch()
+        actions = []
+        for inv in files:
+            doc = {}
+            with open(inv) as f:
+                doc = self.parser.parse_file(f)
+                actions.append({'_index': self.index_name, '_id': doc['id'], '_source': doc})
+        # add to index:
+        result = bulk(es, actions)
