+++ procrustus_indexer/parsers/__init__.py
@@ -1,0 +1,2 @@
+from .parser import Parser
+from .json_parser import JsonParser
