+++ pyproject.toml
@@ -1,0 +1,20 @@
+[project]
+name = "procrustus-indexer"
+version = "2.0.0"
+description = ""
+authors = [
+    {name = "Meindert Kroese",email = "meindert.kroese@di.huc.knaw.nl"},
+    {name = "Jarno Bakker",email = "jarno.bakker@di.huc.knaw.nl"}
+]
+readme = "README.md"
+requires-python = ">=3.13"
+dependencies = [
+    "jmespath (>=1.0.1,<2.0.0)",
+    "elasticsearch (>=8.17.2,<9.0.0)",
+    "pytest (>=8.3.5,<9.0.0)"
+]
+
+
+[build-system]
+requires = ["poetry-core>=2.0.0,<3.0.0"]
+build-backend = "poetry.core.masonry.api"
