+++ tests/test_files/json_files/incomplete.json
@@ -1,0 +1,9 @@
+{
+  "id": 2,
+  "title": "This is a test",
+  "nested": {
+    "object": {
+      "value": 7
+    }
+  }
+}
