+++ tests/test_files/test-json.toml
@@ -1,0 +1,26 @@
+[index]
+full_text="yes"
+record="yes"
+
+[index.input]
+format="json"
+
+[index.id]
+path="jmes:id"
+type="keyword"
+
+[index.facet.title]
+path="jmes:title"
+type="text"
+
+[index.facet.tag]
+path="jmes:tag"
+type="text"
+
+[index.facet.value]
+path="jmes:value"
+type="number"
+
+[index.facet.nested_value]
+path="jmes:nested.object.value"
+type="number"
